third_party/WebKit/LayoutTests/external/wpt/XMLHttpRequest/send-content-type-charset-expected.txt

This is a testharness.js-based test.
FAIL header with invalid MIME type is not changed assert_equals: expected "text; charset=ascii" but got "text; charset=UTF-8"
PASS known charset but bogus header - missing MIME type
PASS bogus charset and bogus header - missing MIME type
PASS Correct text/plain MIME with charset
PASS If no charset= param is given, implementation should not add one - unknown MIME
PASS If no charset= param is given, implementation should not add one - known MIME
PASS charset given but wrong, fix it (unknown MIME, bogus charset)
PASS charset given but wrong, fix it (known MIME, bogus charset)
PASS charset given but wrong, fix it (known MIME, actual charset)
PASS If multiple charset parameters are given, all should be rewritten
FAIL A parameter name should be accompanied by a parameter value assert_equals: expected "text/x-pink-unicorn;x;charset=us-ascii" but got "text/x-pink-unicorn;x;charset=UTF-8"
FAIL Quoted string should be handled correctly assert_equals: expected "text/x-pink-unicorn;x=\" y z u charset=t\";charset=UTF-8;charset=UTF-8" but got "text/x-pink-unicorn;x=\" y z u charset=UTF-8\";charset=\"UTF-8 b\";charset=UTF-8"
PASS No content type set, give MIME and charset
Harness: the test ran to completion.
